requirements.txt
asgiref==3.8.1
backports.zoneinfo==0.2.1
Brotli==1.1.0
click==8.1.8
dj-database-url==2.3.0
Django==4.2.20
djangorestframework==3.15.2
gunicorn==23.0.0
h11==0.14.0
packaging==24.2
psycopg==3.2.6
psycopg-binary==3.2.6
psycopg2==2.9.10
python-dotenv==1.0.1
sqlparse==0.5.3
typing-extensions==4.12.2
uvicorn==0.33.0
whitenoise==6.7.0
